[REF] stock_landed_costs_average: set common class for unit testing
