+++ stock_landed_costs_average/tests/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from . import test_stock_landed_common
 from . import test_basics
 from . import test_stock_landed_costs_average
 from . import test_landed_costs_with_std
